requirements.txt
admin
asgiref
Django
mysqlclient
sqlparse
python-decouple
django-tinymce
Pillow
tzdata==2022.1